requirements.txt
streamlit
reportlab
Pillow
requests
folium
streamlit_folium
